+++ src/main/resources/templates/error/404.hbs
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>체스</title>
+    <link rel="stylesheet" href="/css/index.css">
+</head>
+<body>
+
+<div class="mainContainer">
+    <h2>페이지를 찾을 수 없습니다!</h2>
+    <a href="/">메인으로</a>
+</div>
+
+</body>
+</html>
